requirements.txt
fastapi~=0.108.0
pydantic~=2.5.3
pythondi~=1.2.4
python-dotenv==0.19.2

uuid~=1.30
starlette~=0.32.0.post1
pymongo~=4.6.1
motor~=3.3.2
databases